claude: refactor(kaspa): rename kas_relayer and kas_validator modules
Remove redundant 'kas_' prefix from module names since they're already
inside the dymension-kaspa crate. This improves code clarity without
changing functionality.
