+++ rust/main/chains/dymension-kaspa/src/relayer/withdraw/messages.rs
@@ -2,11 +2,9 @@ use super::hub_to_kaspa::{
     build_withdrawal_pskt, extract_current_anchor, fetch_input_utxos, get_normal_bucket_feerate,
     get_outputs_from_msgs,
 };
-use crate::kas_relayer::withdraw::sweep::{
-    create_inputs_from_sweeping_bundle, create_sweeping_bundle,
-};
 use crate::ops::payload::MessageIDs;
 use crate::ops::withdraw::{filter_pending_withdrawals, WithdrawFXG};
+use crate::relayer::withdraw::sweep::{create_inputs_from_sweeping_bundle, create_sweeping_bundle};
 use dym_kas_core::consts::RELAYER_SIG_OP_COUNT;
 use dym_kas_core::escrow::EscrowPublic;
 use dym_kas_core::pskt::{estimate_mass, PopulatedInput};
