+++ rust/main/chains/dymension-kaspa/src/validator/confirmation.rs
@@ -1,6 +1,6 @@
-use crate::kas_validator::error::ValidationError;
 use crate::ops::confirmation::ConfirmationFXG;
 use crate::ops::payload::{MessageID, MessageIDs};
+use crate::validator::error::ValidationError;
 use dym_kas_api::models::{TxModel, TxOutput};
 use dym_kas_core::api::client::HttpClient;
 use dym_kas_core::finality::is_safe_against_reorg;
